+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 - Add magento errors to paazl_log for cronjobs
+- Bugfix no housenumber on payment step in checkout
 
 ## [1.2.2] - 2017-05-05
 ### Changed